backend/requirements.txt

Flask==2.3.4
Flask-Cors==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==2.3.7
PyJWT==2.8.0
